+++ .env
@@ -1,1 +1,2 @@
+# VITE_BASE_API_URL=https://sport-api-92rv.onrender.com
 VITE_BASE_API_URL=http://localhost:5000
