+++ package.json
@@ -25,7 +25,7 @@
     "react-dom": "^17.0.0 || ^18.2.0",
     "react-dropzone": "^14.2.3",
     "react-material-ui-carousel": "^3.4.2",
-    "react-router-dom": "^6.21.1",
+    "react-router-dom": "^6.22.0",
     "styled-components": "^6.1.8",
     "validator": "^13.11.0"
   },
